LayoutTests/platform/ios/imported/w3c/web-platform-tests/xhr/setrequestheader-case-insensitive-expected.txt


PASS XMLHttpRequest: setRequestHeader() - headers that differ in case
FAIL XMLHttpRequest: setRequestHeader() - headers that differ in case 1 assert_regexp_match: expected object "/content-TYPE/" but got "Host: localhost:8800\nContent-Type: x/x\nConnection: keep-alive\nTHIS-IS-A-TEST: 1, 2\nAccept: */*\nUser-Agent: Mozilla/5.0 (iPhone; CPU iPhone OS 18_0 like Mac OS X) AppleWebKit/605.1.15 (KHTML, like Gecko) Mobile/15E148\nReferer: http://localhost:8800/xhr/setrequestheader-case-insensitive.htm\nAccept-Language: en-US,en;q=0.9\nAccept-Encoding: gzip, deflate\n\n"
